add suspense boundary to audits route
